Move util function to flutter directory
