+++ packages/devtools_app/lib/src/logging/flutter/logging_screen.dart
@@ -14,6 +14,7 @@ import '../../flutter/screen.dart';
 import '../../flutter/split.dart';
 import '../../flutter/table.dart';
 import '../../flutter/theme.dart';
+import '../../flutter/utils.dart';
 import '../../globals.dart';
 import '../../table_data.dart';
 import '../../ui/flutter/service_extension_widgets.dart';
@@ -362,6 +363,7 @@ class _MessageColumn extends LogMessageColumn
         ],
       );
     } else if (data.kind == 'stdout') {
+      // TODO(helin24): Move this to logging controller when dart:html is removed.
       const Color color = Color.fromARGB(0xff, 0x00, 0x91, 0xea);
 
       final RichText richText = RichText(
